Updated example files
