+++ examples/CFLP/model.jl
@@ -119,7 +119,6 @@ function model_definition(problem_params::DynamicSDDiP.ProblemParams, scenario_t
 
         # Add stage objective
         scaling_factor = 1/1000
-        # scaling_factor = 1
         SDDP.@stageobjective(subproblem, scaling_factor * sum(f[i] * y[i] for i in 1:number_of_facilities) + scaling_factor * sum(d[j][i] * x[i,j] for j in 1:number_of_customers for i in 1:number_of_facilities))
 
         # PARAMETERIZE THE RANDOM VARIABLES
