+++ examples/Regularization_Paper_Illustrative/Reg_Paper_Ex_A.jl
@@ -8,9 +8,7 @@ using JuMP
 using SDDP
 using DynamicSDDiP
 using Revise
-#using Gurobi
-using GAMS
-#using SCIP
+using Gurobi
 using Infiltrator
 
 
@@ -92,18 +90,15 @@ function model_config()
     )
 
     # Define solvers to be used
-    applied_solvers = DynamicSDDiP.AppliedSolvers(
-        LP = "Gurobi",
-        MILP = "Gurobi",
-        MIQCP = "Gurobi",
-        MINLP = "SCIP",
-        NLP = "Gurobi",
-        Lagrange = "Gurobi",
-    )
+    applied_solvers = DynamicSDDiP.AppliedSolvers()
+
+    # Define problem params
+    problem_params = DynamicSDDiP.ProblemParams(1, 2)
 
     # Start model with used configuration
     model_starter(
         algo_params,
+        problem_params,
         applied_solvers,
     )
 end
@@ -111,6 +106,7 @@ end
 
 function model_starter(
     algo_params::DynamicSDDiP.AlgoParams = DynamicSDDiP.AlgoParams(),
+    problem_params::DynamicSDDiP.ProblemParams = DynamicSDDiP.ProblemParams(2, 1),
     applied_solvers::DynamicSDDiP.AppliedSolvers = DynamicSDDiP.AppliedSolvers(),
     )
 
@@ -122,7 +118,7 @@ function model_starter(
     ############################################################################
     # SOLVE MODEL
     ############################################################################
-    DynamicSDDiP.solve(model, algo_params, applied_solvers)
+    DynamicSDDiP.solve(model, algo_params, applied_solvers, problem_params)
 end
 
 
@@ -133,7 +129,7 @@ function model_definition()
     model = SDDP.LinearPolicyGraph(
         stages = number_of_stages,
         lower_bound = 0.0,
-        optimizer = GAMS.Optimizer,
+        optimizer = Gurobi.Optimizer,
         sense = :Min
     ) do subproblem, t
