+++ examples/Regularization_Paper_Illustrative/Reg_Paper_Ex_C.jl
@@ -285,7 +285,11 @@ model_starter(6/5, true, 10.0, DynamicSDDiP.L₁(), DynamicSDDiP.L₁(), Dynamic
 # All ALD runs with only ValueBound.
 # Second batch of tests with backward regularization (sigma = rho used in augmented Lagrangian dual), but with using the primal_obj bound from the unregularized case.  
 # Note that for the augmented Lagrangian dual, we always use the L₁-norm.
+
+ReLU comparison
+model_starter(1.249, true, 5.0, DynamicSDDiP.L₁(), DynamicSDDiP.L₁(), DynamicSDDiP.StateSpaceCopy(), DynamicSDDiP.ConvexHullCopy(), true, 8, false) 
+
 """
 
-model_starter(1.249, true, 5.0, DynamicSDDiP.L₁(), DynamicSDDiP.L₁(), DynamicSDDiP.StateSpaceCopy(), DynamicSDDiP.ConvexHullCopy(), true, 8, false)   # ReLU comparison
+model_starter(6/5, false, Inf, nothing, nothing, DynamicSDDiP.StateSpaceCopy(), DynamicSDDiP.StateSpaceCopy(), false, 0, false)
